pyproject.toml
[project]
name = "agentsmcp"
version = "1.0.0"
description = "CLI-driven MCP agent system with extensible RAG pipeline"
requires-python = ">=3.10"
dependencies = [
    # Core framework & tooling
    "click",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0,<3.0",
    "pydantic-settings",
    "pyyaml",
    "httpx",
    "structlog",
    "rich",
    "openai-agents>=0.2.9,<0.3.0",
    "openai>=1.0.0,<2.0.0",
    "python-json-logger",
    # Metrics/discovery/security moved to extras to keep minimal install light
    "prompt_toolkit>=3.0.0",
]

[project.scripts]
agentsmcp = "agentsmcp.cli_minimal:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "bandit",
    "pip-audit",
]
rag = [
    "sentence-transformers",
    "numpy",
]
ui = [
    # UI relies on rich (core); keep extra placeholder for clarity
]
cost = [
    # cost features currently have no extra third-party deps
]
discovery = [
    "aiohttp>=3.8.0,<4.0.0",
    "websockets>=11.0.0,<12.0.0",
    "jsonschema>=4.0.0,<5.0.0",
]
security = [
    "cryptography>=41.0.0,<42.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
]
metrics = [
    "prometheus-fastapi-instrumentator",
    "opentelemetry-api>=1.20.0,<2.0.0",
    "opentelemetry-sdk>=1.20.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.41.0,<1.0.0",
    "opentelemetry-exporter-jaeger-thrift>=1.20.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0,<2.0.0",
    "opentelemetry-exporter-prometheus>=1.12.0,<2.0.0",
    "prometheus-client>=0.17.0,<1.0.0",
]
bench = [
    # placeholder for future benchmarking deps
]

[tool.pytest.ini_options]
addopts = "--cov=agentsmcp --cov-fail-under=80"

[tool.ruff]
line-length = 88
exclude = [
    "src/agentsmcp/orchestration",
    "src/agentsmcp/ui",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "E402", "I001"]

[tool.bandit]
skips = ["B101"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["agentsmcp*"]
